pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docker-compose-doctor"
version = "0.1.0"
description = "Diagnose and recover from Docker Compose volume drift and broken bind mount paths"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Imhotep Systems", email = "hello@imhotepsystems.com" }
]
keywords = ["docker", "docker-compose", "devtools", "volumes", "diagnostics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "rich>=13.0",
    "pyyaml>=6.0",
    "docker>=6.0",
    "thefuzz>=0.20",
]

[project.scripts]
docker-compose-doctor = "docker_compose_doctor.cli:main"

[project.urls]
Homepage = "https://imhotepsystems.com/tools/docker-compose-doctor"
Repository = "https://github.com/imhotep-systems/docker-compose-doctor"
Issues = "https://github.com/imhotep-systems/docker-compose-doctor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docker_compose_doctor"]
